requirements.txt
python-dotenv ~= 0.18.0

discord.py ~= 1.7.2
discord_components~=1.1.4

colorama~=0.4.4
termcolor~=1.1.0
psutil~=5.8.0
py-cpuinfo~=8.0.0

aiosqlite~=0.17.0
embed-templator>=1.0.3